Use TxComposer in register multisig form
