+++ src/locales/en/common.json
@@ -177,7 +177,6 @@
   "German": "German",
   "Go Back": "Go Back",
   "Go back": "Go back",
-  "Go to confirmation": "Go to confirmation",
   "Go to dashboard": "Go to dashboard",
   "Great! You have an account now": "Great! You have an account now",
   "Height": "Height",
